Converted schedule visualizer to component
