+++ rmf_visualization_schedule/package.xml
@@ -14,6 +14,7 @@
   <depend>rmf_traffic_ros2</depend>
   <depend>rmf_traffic_msgs</depend>
   <depend>rclcpp</depend>
+  <depend>rclcpp_components</depend>
   <depend>libwebsocketpp-dev</depend>
   <depend>geometry_msgs</depend>
   <depend>visualization_msgs</depend>
